refactor ♻️: Update subtitle readability constraints in constant.py
This change updates the subtitle readability constraints in the `constant.py` file by removing the reference to `AlignedResult`. The update is made to align with the new naming conventions for timestamps, improving consistency and clarity in the codebase.
